Progress commit.
Includes a bunch of generated CSS using new gulp watcher processing for domains.

Also includes:

Added a wrapper around the include of syntric-apps/syntric-app.php in functions.php
that checks if ACF plugin is active.  Was causing Syntric theme to error
as ACF is a dependency.

Updated gulpfile.js so that it now generates domain-centric names for CSS files.